requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
scipy==1.11.4
plotly==5.18.0
PyYAML==6.0.1
openpyxl==3.1.2
tqdm==4.66.1
